refactor: actions 위치 변경
