+++ src/app/sample/actions.ts
@@ -1,0 +1,9 @@
+'use server';
+
+export async function componentTestAction(currentState, formData: FormData) {
+  const data = Object.fromEntries(formData);
+  console.log('data', data);
+  return {
+    message: 'sample action success',
+  };
+}
